[SYCLomatic] Refactor CubRule
