+++ clang/lib/DPCT/APINamesCUB.inc
@@ -11,39 +11,45 @@ CONDITIONAL_FACTORY_ENTRY(
     CheckCubRedundantFunctionCall(),
     REMOVE_API_FACTORY_ENTRY("cub::DeviceReduce::Sum"),
     HEADER_INSERT_FACTORY(
-        HeaderType::HT_DPL_Algorithm_Execution,
+        HeaderType::HT_DPL_Execution,
+        HEADER_INSERT_FACTORY(
+        HeaderType::HT_DPL_Algorithm,
         REMOVE_CUB_TEMP_STORAGE_FACTORY(MEMBER_CALL_FACTORY_ENTRY(
             "cub::DeviceReduce::Sum", QUEUESTR, false, "fill", ARG(3),
             CALL("oneapi::dpl::reduce",
                  CALL("oneapi::dpl::execution::device_policy", QUEUESTR),
                  ARG(2), BO(BinaryOperatorKind::BO_Add, ARG(2), ARG(4))),
-            LITERAL("1")))))
+            LITERAL("1"))))))
 
 // cub::DeviceScan::ExclusiveSum
 CONDITIONAL_FACTORY_ENTRY(
     CheckCubRedundantFunctionCall(),
     REMOVE_API_FACTORY_ENTRY("cub::DeviceScan::ExclusiveSum"),
     HEADER_INSERT_FACTORY(
-        HeaderType::HT_DPL_Algorithm_Execution,
+        HeaderType::HT_DPL_Execution,
+        HEADER_INSERT_FACTORY(
+        HeaderType::HT_DPL_Algorithm,
         REMOVE_CUB_TEMP_STORAGE_FACTORY(CALL_FACTORY_ENTRY(
             "cub::DeviceScan::ExclusiveSum",
             CALL("oneapi::dpl::exclusive_scan",
                  CALL("oneapi::dpl::execution::device_policy", QUEUESTR),
                  ARG(2), BO(BinaryOperatorKind::BO_Add, ARG(2), ARG(4)), ARG(3),
-                 LITERAL("0"))))))
+                 LITERAL("0")))))))
 
 // cub::DeviceScan::InclusiveSum
 CONDITIONAL_FACTORY_ENTRY(
     CheckCubRedundantFunctionCall(),
     REMOVE_API_FACTORY_ENTRY("cub::DeviceScan::InclusiveSum"),
     HEADER_INSERT_FACTORY(
-        HeaderType::HT_DPL_Algorithm_Execution,
+        HeaderType::HT_DPL_Execution,
+        HEADER_INSERT_FACTORY(
+        HeaderType::HT_DPL_Algorithm,
         REMOVE_CUB_TEMP_STORAGE_FACTORY(CALL_FACTORY_ENTRY(
             "cub::DeviceScan::InclusiveSum",
             CALL("oneapi::dpl::inclusive_scan",
                  CALL("oneapi::dpl::execution::device_policy", QUEUESTR),
                  ARG(2), BO(BinaryOperatorKind::BO_Add, ARG(2), ARG(4)),
-                 ARG(3))))))
+                 ARG(3)))))))
 
 // cub::DeviceSelect::Flagged
 CONDITIONAL_FACTORY_ENTRY(
