+++ clang/test/dpct/cub/devicelevel/device_exclusive_sum.cu
@@ -57,7 +57,6 @@ void test_1() {
 // device_in = sycl::malloc_device<int>(n, q_ct1);
 // device_out = sycl::malloc_device<int>(n, q_ct1);
 // q_ct1.memcpy(device_in, (void *)host_in, sizeof(host_in)).wait();
-// DPCT1026:{{.*}}
 // DPCT1027:{{.*}}
 // 0, 0;
 // oneapi::dpl::exclusive_scan(oneapi::dpl::execution::device_policy(q_ct1), device_in, device_in + n, device_out, 0);
