+++ clang/test/dpct/cub/devicelevel/device_scan_flagged.cu
@@ -68,7 +68,6 @@ void test_1() {
 // CHECK: device_select_num = sycl::malloc_device<int>(1, q_ct1);
 // CHECK: q_ct1.memcpy(device_in, (void *)host_in, sizeof(host_in));
 // CHECK: q_ct1.memcpy(device_flagged, host_flagged, sizeof(host_flagged)).wait();
-// CHECK: DPCT1026:{{.*}}
 // CHECK: DPCT1027:{{.*}}
 // CHECK: 0, 0;
 // CHECK: q_ct1.fill(device_select_num, dpct::copy_if(oneapi::dpl::execution::device_policy(q_ct1), device_in, device_in + n, device_flagged, device_out, [](const auto &t) -> bool { return t; }) - device_out, 1);
